fix: updated deleteMemo method to include category parameter
- 메모 삭제 시 AI 분류 실패 카테고리임이 확인되면 파이어스토어에서의 삭제 동작을 막았음.
- 실패 시 모든 처리는 오직 로컬에서만 이뤄지고 분류가 성공했을 경우에만 서버로 전파할 수 있도록 하는 기초 토대 마련

+++ lib/data/repositories/memo/memo_repository_remote.dart
@@ -64,6 +64,9 @@ class MemoRepositoryRemote extends MemoRepository {
 
       _hiveService.saveMemo(analyzedMemo, uuid);
       debugPrint('✅ 하이브 저장 완료');
+      if (analyzedMemo.category == 'AI분류 실패') {
+        return "AI분류 실패";
+      }
       _firestoreService.saveMemo(analyzedMemo, uuid);
       debugPrint('✅ 파이어스토어 저장 완료');
       return null;
@@ -113,11 +116,15 @@ class MemoRepositoryRemote extends MemoRepository {
   }
 
   @override
-  Future<void> deleteMemo(String memoId) async {
+  Future<void> deleteMemo(String memoId, String category) async {
     try {
-      await _firestoreService.deleteMemo(memoId);
       _hiveService.deleteMemo(memoId);
-      debugPrint('✅ 메모 삭제 완료');
+      debugPrint('✅ 하이브에서 삭제 완료');
+      if (category == 'AI분류 실패') {
+        return;
+      }
+      await _firestoreService.deleteMemo(memoId);
+      debugPrint('✅ firestore에서 삭제 완료');
     } catch (e) {
       debugPrint('❌ 메모 삭제 실패 in [deleteMemo method] in [memo_repository_remote]: $e');
       rethrow;
